Split store and retrieve into dedicated screens
